add queue, mc mr belum dibikin

+++ src/Expression/Binary/AddExpression.cpp
@@ -6,6 +6,6 @@
 AddExpression::AddExpression(Expression *x, Expression *y): BinaryExpression(x, y) {
 }
 
-float AddExpression::solve() {
+double AddExpression::solve() {
 	return (x->solve() + y->solve());
 }
