+++ src/Expression/Binary/DivideExpression.cpp
@@ -6,6 +6,6 @@
 DivideExpression::DivideExpression(Expression *x, Expression *y): BinaryExpression(x, y) {
 }
 
-float DivideExpression::solve() {
+double DivideExpression::solve() {
 	return (x->solve() / y->solve());
 }
